requirements.txt
flask==2.3.3
werkzeug==3.0.1
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0